add mailhog support + document performance issues with docker for mac
Added mailhog support so that both the test pipelines themselves can leverage it
during test runs, but also for the benefit of the pipeline itself to test its
email functionalities (e.g. email notifications at end of Jenkins build).

Also added a note in the documentation on performance issues with Docker for Mac
because of the host mounts. No issues on Linux, but on Docker for Mac, this
causes I/O to be very slow, which is an issue for I/O intensive tasks (e.g.
builds in Jenkins).

You can either:

 * use your own solution to sync with the container (e.g.
   docker-sync, maybe),
 * or use more complicated mount solutions,
 * or remove all host-mounts entirely if you don't need easy access and
    persistence,
 * or configure the service so that some internal folders reside in locations
   where they are NOT part of the host-mounted folder.

e.g. typically, your jenkins workspace, which you may want to simply put under
/tmp, whereas the default JENKINS_HOME - /var/jenkins_home - is host-mounted and
thus very slow for I/Os.